feat(writer): metagraph writer
Adds a _MetaGraphWriter, exporting the encountered taxonomy as a graph, along with properties usage statistics.
Each type becomes a node, and number of instances encountered during the processing is added as a property.
Each encountered property value is added in a list attached as a property, and its cardinality is also attached as a property.

Allows to export in NetworkX-supported formats: cytoscape, dot, graphml, gefx, gml or pajek.