fix(ontology): connect dangling user extension classes in the NX graph
Classes added to the ontology by the user (with `is_a`)
were added to the NX *graph* of ontologies,
but were not guaranteed to be linked to the *tree*.

First iterations of the NX graph construction from ontologies' taxonomy
does guarantee that no dangling class exists, but when adding
user extensions, this was not guaranteed anymore.

This commit adds the Ontology._connect_dangling_classes function,
which arbitrary link dangling classes to the user-defined root class
of the head ontology.

It does so with a warning, to maintain backward compatibility,
but this should really be an error (in future versions?), since
it most probably come from an error on the user's schema.